refactor: restructure shared schemas
