+++ schemas/shared/link.ts
